[IMP] core: move monkey patches to a separate package
In order to make the `odoo` python package a native-namespace, in this
change we want to isolate various monkey-patches found inside
`__init__.py` into dedicates packages inside `_monkeypatching`.
We also move the patches that should be deprecated so we can remove them
easily in the future.

Summary of changes:
- evented is moved to `evented.py` is still patched first when when
  starting odoo
- it's just followed by setting timezone to UTC
- patching of codecs (and language-related stuff) and mimetypes is also
  gathered, there is no reason to have it inside `__init__` or `http`
- patching pytz, literal_eval, num2words, stdnum, werkzeug are isolated
- pypdf2 patch is done when importing the library in `odoo.tools.pdf`
- safe_eval makes sure that pytz is patched before exposing it
- xrld.xlsx patch is done when importing `base_import` as it's the main
  usage of this package, and xlrd is deprecated in favor of openpyxl
