+++ addons/base_vat/tests/test_validate_ruc.py
@@ -1,6 +1,6 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 from odoo.tests.common import TransactionCase, tagged
-from odoo.tools._monkeypatches import new_get_soap_client
+from odoo._monkeypatches.stdnum import new_get_soap_client
 from odoo.exceptions import ValidationError
 from unittest.mock import patch
