+++ odoo/_monkeypatches/num2words.py
@@ -1,9 +1,12 @@
 import decimal
+import logging
 import math
 import re
 from collections import OrderedDict
 from decimal import ROUND_HALF_UP, Decimal
 from math import floor
+
+from odoo import MIN_PY_VERSION
 
 # The following section of the code is used to monkey patch
 # the Arabic class of num2words package as there are some problems
@@ -966,3 +969,16 @@ class NumberToWords_BG(Num2Word_Base):
         ret = self._discard_empties(ret)
         ret.reverse()
         return ret_minus + ''.join(ret)
+
+
+def patch_num2words():
+    try:
+        import num2words  # noqa: PLC0415
+    except ImportError:
+        _logger = logging.getLogger(__name__)
+        _logger.warning("num2words is not available, Arabic number to words conversion will not work")
+        return
+    if MIN_PY_VERSION >= (3, 12):
+        raise RuntimeError("The num2words monkey patch is obsolete. Bump the version of the library to the latest available in the official package repository, if it hasn't already been done, and remove the patch.")
+    num2words.CONVERTER_CLASSES["ar"] = Num2Word_AR_Fixed()
+    num2words.CONVERTER_CLASSES["bg"] = NumberToWords_BG()
