+++ odoo/_monkeypatches/pytz.py
@@ -119,14 +119,13 @@ _tz_mapping = {
     "Zulu": "Etc/UTC",
 }
 
-
 original_pytz_timezone = pytz.timezone
 
 
-def timezone(name):
-    if name not in pytz.all_timezones_set and name in _tz_mapping:
-        name = _tz_mapping[name]
-    return original_pytz_timezone(name)
+def patch_pytz():
+    def timezone(name):
+        if name not in pytz.all_timezones_set and name in _tz_mapping:
+            name = _tz_mapping[name]
+        return original_pytz_timezone(name)
 
-
-pytz.timezone = timezone
+    pytz.timezone = timezone
